+++ core/src/proto/blockchain.rs
@@ -36,7 +36,7 @@ pub enum BlockFull {
         block: BlockData,
         proof: Bytes,
         queue_diff: Bytes,
-        save_utime: u32,
+        save_utime: u64,
     },
     #[tl(id = "blockchain.blockFull.notFound")]
     NotFound,
